node/approval-voting: continue progress on full subsystem tests
This commit expands on the prior commit to attempt to have a first
end-to-end full subsystem test for approval voting. In future commits
the body of this test will be broken out into helper functions to enable
multiple modular tests without duplicating code. In this diff the goal was
to get the first strides for such full subsystem tests working.